fix: allow non-visual ranges
Now users can type `1,100LspCodeAction` and have it actually work. The
downside is that this only works line-wise. I'll have to add equivalent
visual mode mappings for char-wise actions.

Also, the current implementation relies on a Big Hack™: I use `99999999`
as the end column number because that's the easiest way I found to get
`vim.lsp.util.make_given_range_params()` to give the correct index for
the last character on a given line.